MMS: Fix make Name mode searching better
The name search mode is use name to find phone number, then use this number
to find thread, and display all the messages in this thread. But if the
thread have more than one number, it can't match the name's number.

Find all the threads which contains the name's number.

CRs-Fixed: 588037

Change-Id: I822f96dc342ec6982a09a60cf0d7bcaa6d101613