Mms: Fix issue that message will crash when sending mms and can not download mms.
When sending mms on CT, do not get the outputstream of the attachment
by uri due to use incorrect methord to compare the path, and finally
it occurs null pointer exception.

Use canonical path to determin whether two paths point to the same
file, because it will change symbolic link to canonical paths when
comparing the paths.

Change-Id: Iab88744b664748bc590b0df488b6c9f4c5529ebf
CRs-Fixed: 637113