TelephonyProvider:Fix Messaging crash after upgrading from KK
"phone_id" column not found exception happened due to database
inconsistant. The message database version on KK is 58. When OTA
KK to L, the message database upgrade starts from version 59 to
latest(61). However, the "phone_id" column isn't added because
it is lies in the version 58 upgrade method.

Check the "phone_id" column, if it doesn't exist when upgrade
to version 59, this means the database may come from a KK
platform. We need add the "phone_id" column.

Change-Id: Ia6dcd5811ea55ee2752224fe9e4186c49565b021
CRs-Fixed: 783252